refactor: update forget interaction methods

+++ src/Traits/InteractionRelationship.php
@@ -67,7 +67,7 @@ trait InteractionRelationship
      *
      * @return static
      */
-    public function forgetInteractions(string $interactionType): static
+    public function forgetInteractionsOfType(string $interactionType): static
     {
         $this->likes()->where('type', $interactionType)->delete();
 
@@ -77,10 +77,16 @@ trait InteractionRelationship
     /**
      * Check and forget all recorded interactions.
      *
+     * @param  null|string  $interactionType
+     *
      * @return static
      */
-    public function forgetAllInteractions(): static
+    public function forgetInteractions(?string $interactionType = null): static
     {
+        if (in_array($interactionType, InteractionTypeEnum::getValuesAsStrings())) {
+            return $this->forgetInteractionsOfType($interactionType);
+        }
+
         $this->likes()->delete();
 
         return $this;
